+++ tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "@tsconfig/node18-strictest/tsconfig.json",
+  "include": [
+    "src"
+  ],
+  "compilerOptions": {
+    "outDir": "build",
+    "declaration": true
+  }
+}
